Fixes problem w/ custom ontology set changes
The cache was not being invalidated properly when users were updating
their custom ontology set. Also, the list of ontologies to choose from
was limited to the set the user had previously picked, making changes
impossible. There are two fixes here. The first invalidates the cache
on the REST and UI side appropriately. The second invokes a new
parameter ‘ignore_custom_ontologies’ that allows us to get an
unfiltered list of ontologies where necessary.